requirements.txt
#Numerical package 
numpy>=2.1.0

# Core AI/ML Framework
langchain>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.2.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.2.0


# Web Interface
streamlit>=1.30.0

# API Gateway
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Document Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
python-pptx>=0.6.0

# Telegram Bot
python-telegram-bot>=21.0

# Vector Database
chromadb>=0.5.0
sentence-transformers>=2.5.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0